move graphql select into shared function
