+++ src/packages/admin-ui-components/src/detail-panel/graphql.ts
@@ -1,6 +1,6 @@
 import { gql } from '@apollo/client';
 import pluralize from 'pluralize';
-import { Entity } from '../utils';
+import { Entity, generateGqlSelectForEntityFields } from '../utils';
 
 export const generateUpdateEntityMutation = (
 	entity: Entity,
@@ -9,16 +9,7 @@ export const generateUpdateEntityMutation = (
     mutation updateEntity ($data: ${entity.name}CreateOrUpdateInput!){
       update${entity.name} (data: $data) {
         id
-        ${entity.fields
-					.map((field) => {
-						if (field.relationshipType) {
-							const relatedEntity = entityByType(field.type);
-							return `${field.name} { id ${relatedEntity.summaryField || ''} }`;
-						} else {
-							return field.name;
-						}
-					})
-					.join(' ')}
+        ${generateGqlSelectForEntityFields(entity, entityByType)}
       }
     }
   `;
@@ -30,16 +21,7 @@ export const generateCreateEntityMutation = (
     mutation createEntity ($data: ${entity.name}InsertInput!){
       create${entity.name} (data: $data) {
         id
-        ${entity.fields
-					.map((field) => {
-						if (field.relationshipType) {
-							const relatedEntity = entityByType(field.type);
-							return `${field.name} { id ${relatedEntity.summaryField || ''} }`;
-						} else {
-							return field.name;
-						}
-					})
-					.join(' ')}
+        ${generateGqlSelectForEntityFields(entity, entityByType)}
       }
     }
   `;
