+++ src/packages/admin-ui/src/pages/list/graphql.ts
@@ -1,5 +1,5 @@
 import { gql } from '@apollo/client';
-import { Entity } from '@exogee/graphweaver-admin-ui-components';
+import { Entity, generateGqlSelectForEntityFields } from '@exogee/graphweaver-admin-ui-components';
 import pluralize from 'pluralize';
 
 export const queryForEntityPage = (entityName: string, entityByType: (type: string) => Entity) => {
@@ -10,16 +10,7 @@ export const queryForEntityPage = (entityName: string, entityByType: (type: stri
 	return gql`
 		query AdminUIListPage($filter: ${pluralName}ListFilter, $pagination: ${pluralName}PaginationInput) {
 			result: ${queryName}(filter: $filter, pagination: $pagination) {
-				${entity.fields
-					.map((field) => {
-						if (field.relationshipType) {
-							const relatedEntity = entityByType(field.type);
-							return `${field.name} { id ${relatedEntity.summaryField || ''} }`;
-						} else {
-							return field.name;
-						}
-					})
-					.join(' ')}
+				${generateGqlSelectForEntityFields(entity, entityByType)}
 			}
 		}
 	`;
